requirements.txt
asgiref==3.8.1
Django==5.1.2
django-admin-thumbnails==0.2.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==2.9.0
six==1.16.0
sqlparse==0.5.1
